backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
supabase==2.7.4
openai==1.47.0
yfinance==0.2.43
requests==2.32.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.9
email-validator==2.2.0
